Fix for #63
Moment’s `range.contains` function take timestamps into account when
computing if a date is within the given range. `CalendarDate` is
creating a date at midnight on the day in question, which will never be
within the starting range. By setting the day’s time to “one minute
past midnight”, this allows the date comparison to behave a bit more
logically.