Don't rely on moment-range overlaps method
The moment-range `overlaps` method doesn't consider ranges which 'touch' on equal start/end moments to be overlapping. This breaks highlighting and selection display across month boundaries.

Instead, explicitly test if the first range contains the start and end days of the second range, as well as checking for overlaps.

Oops again